third_party/blink/web_tests/platform/webview/external/wpt/animation-worklet/idlharness.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] WorkletAnimation must be primary interface of new WorkletAnimation("name")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: WorkletAnimation is not defined"
[FAIL] Stringification of new WorkletAnimation("name")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: WorkletAnimation is not defined"
[FAIL] WorkletAnimation interface: new WorkletAnimation("name") must not have property "undefined"
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: WorkletAnimation is not defined"
[FAIL] WorkletAnimation interface: new WorkletAnimation("name") must not have property "animatorName"
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: WorkletAnimation is not defined"
[FAIL] CSS namespace: operation escape(CSSOMString)
  Cannot read properties of undefined (reading 'hasOwnProperty')
[FAIL] CSS namespace: attribute animationWorklet
  Cannot read properties of undefined (reading 'hasOwnProperty')
Harness: the test ran to completion.
